Improved accuracy for OCR; added file processing
1.Added in a series of morphological transformations that allowed pytess
accuracy to improve greatly. Gaussian blur made the biggest difference
in improving the accuracy with number recognition.

2.Automated file processing was completed and ready for export into SQL
database.

NOTE: Unit Price for items had to be exported separately as pytess. is
unable to read single hyphens despite my best efforts and I figure it
would be easier to re-add that column on the database end.